+++ lib/output.sh
@@ -4,6 +4,7 @@
 # however, it helps Shellcheck realise the options under which these functions will run.
 set -euo pipefail
 
+ANSI_BLUE='\033[1;34m'
 ANSI_RED='\033[1;31m'
 ANSI_YELLOW='\033[1;33m'
 ANSI_RESET='\033[0m'
@@ -26,6 +27,24 @@ function output::step() {
 # ```
 function output::indent() {
 	sed --unbuffered "s/^/       /"
+}
+
+# Output a styled multi-line notice message to stderr.
+#
+# Usage:
+# ```
+# output::notice <<-EOF
+# 	Note: The note summary.
+#
+# 	Detailed description.
+# EOF
+# ```
+function output::notice() {
+	echo >&2
+	while IFS= read -r line; do
+		echo -e "${ANSI_BLUE} !     ${line}${ANSI_RESET}" >&2
+	done
+	echo >&2
 }
 
 # Output a styled multi-line warning message to stderr.