+++ spec/hatchet/poetry_spec.rb
@@ -245,6 +245,28 @@ RSpec.describe 'Poetry support' do
       app.deploy do |app|
         expect(clean_output(app.output)).to include(<<~OUTPUT)
           remote: -----> Python app detected
+          remote: 
+          remote:  !     Warning: Multiple Python package manager files were found.
+          remote:  !     
+          remote:  !     Exactly one package manager file should be present in your app's
+          remote:  !     source code, however, several were found:
+          remote:  !     
+          remote:  !     requirements.txt (pip)
+          remote:  !     poetry.lock (Poetry)
+          remote:  !     
+          remote:  !     For now, we will build your app using the first package manager
+          remote:  !     listed above, however, in the future this warning will become
+          remote:  !     an error.
+          remote:  !     
+          remote:  !     Decide which package manager you want to use with your app, and
+          remote:  !     then delete the file(s) and any config from the others.
+          remote: 
+          remote: 
+          remote:  !     Note: We recently added support for the package manager Poetry.
+          remote:  !     If you are using a third-party Poetry buildpack you must remove
+          remote:  !     it, otherwise the requirements.txt file it generates will cause
+          remote:  !     the warning above.
+          remote: 
           remote: -----> Using Python #{DEFAULT_PYTHON_MAJOR_VERSION} specified in .python-version
           remote: -----> Installing Python #{LATEST_PYTHON_3_12}
           remote: -----> Installing pip #{PIP_VERSION}, setuptools #{SETUPTOOLS_VERSION} and wheel #{WHEEL_VERSION}